Second milestone: match the output of the second layer to that of torch's (SpatialBatchNormalization)
The dimension of the input is:
  (batchSize, nInputPlane, inputWidth, inputHeight)
The tricky part is to loop through every of nInputPlane, since it's
the second dimension, things are no longer continuous. However,
once we visualize the way high-dimensional data is stored in memory,
it's actually an easy task.

Also note that this layer carries its own weight and bias.